+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.11](https://github.com/alirezahematidev/jsx-marker/compare/v0.0.10...v0.0.11) (2024-07-18)
+
+
+### Features
+
+* add rest of matches to matcher callback ([bd1bd48](https://github.com/alirezahematidev/jsx-marker/commit/bd1bd48e1f36747481f76d18cf63d7190211d1cf))
+* custom keys now is typed-safe ([29b959b](https://github.com/alirezahematidev/jsx-marker/commit/29b959b72894fc7565308bd8fa06598203f93279))
+
+
+
 ## [0.0.10](https://github.com/alirezahematidev/jsx-marker/compare/v0.0.9...v0.0.10) (2024-04-07)